Combine the two sets of tests for CacheDescriptor
